Add Patch spec and implementation
Patch will be used to keep track of changes that should be applied to
the rom, before actually writing them. It supports bitwise operations
that can be used to combine changes made at different times. This
instance needs to eventually be used to pass the final set of patches
to the RomWriter, though that has not yet been implemented.

The RomWriter is said to expect a Patch in the JSDoc, but this is as
of yet incorrect, and the spec needs to be modified for the RomWriter
to be able to accept a Patch instance.