Make RomWriter take a Patch instance
Spec and implementation have been updated to require the RomWriter to
take a Patch instance rather than an array of patches. The format of
the actual patches has additionally changed from an array of arrays
containing patch data to an object of which both the keys and values
contain patch data.

To summarise the change, this is how it was before:

  [[address, [data, ...]], ...]

And this is how it is now:

  { address: [data,...] }

This was already implemented as part of the Patch class in an earlier
commit, but up until now RomWriter didn't take a Patch instance, so
the implementation didn't need to be there yet. With these classes now
working together, RomWriter is required to adapt to the new format.