Fix a bug where the addresses are calculated incorrectly
I had the domain table turned around (i.e. treating columns like rows
and rows like columns) which naturally caused issues. I also updated
the babel build script so that it actually copies the json files into
the lib folder, which is fairly important!

This is the first build in which I actually got a patch to work just
right; Using the CustomiZ3-CLI project I was able to patch a rom and
make the changes I wanted to make. So finally, I can rest... for a
little bit.