Adjusted the addresses and how data is retrieved for whirlpools
Whirlpool names map to the location they go *to*, not where you
come from if you go through them. This means that for example the
setting a whirlpool to the 'hylia' whirlpool would result in it
ending up by the falls. This change addresses that and makes sure
that if you specify a location, it's where the whirlpool actually
leads to.